+++ src/os/kernel/drivers/disk/ata/ata.h
@@ -19,4 +19,4 @@ typedef struct{
 
 disk_device_t* ATA_identify(uint32_t chann, uint32_t sub);
 bool ATA_read(disk_device_t *disk, uint64_t lba, uint16_t *buff);
-bool ATA_write(disk_device_t *disk, uint32_t lba, uint16_t *buff);
+bool ATA_write(disk_device_t *disk, uint64_t lba, uint16_t *buff);
